+++ tests/bases/test_collections.py
@@ -180,7 +180,11 @@ def test_metric_collection_prefix_postfix_args(prefix, postfix):
     for k in new_metric_collection.keys(keep_base=True):
         assert 'new_prefix_' not in k
 
-    repr = f'MetricCollection(\n  prefix=new_prefix_,\n  postfix={postfix},\n  (DummyMetricSum): DummyMetricSum()\n  (DummyMetricDiff): DummyMetricDiff()\n)'
+    repr = 'MetricCollection(' \
+           '\n  prefix=new_prefix_,' \
+           f'\n  postfix={postfix},' \
+           '\n  (DummyMetricSum): DummyMetricSum()' \
+           '\n  (DummyMetricDiff): DummyMetricDiff()\n)'
     assert new_metric_collection.__repr__() == repr
 
     new_metric_collection = new_metric_collection.clone(postfix='_new_postfix')